Split signature requests watcher in two files
